crucio-responsive-one-page-wordpress-theme-
Crucio is a Responsive One Page WordPress Theme for freelancers, designers, photographers, illustrators, visual artists, creative studios or digital agencies.

It’s based on Bootstrap framework and HTML5/CSS3 technologies, with a charming design, clean and modern, fully responsive and customizable.

Crucio is the perfect solution for any type of business, specially smart companies or startup with a very modern style, clean and elegant.

Quick and easy setup to create a portfolio and show your works and services. The Theme is optimized for smartphone and tablets, it provides a clear documentation and the code is clean and well-organized.

This Theme uses Parallax effect and content animations, so you can create an awesome website. Your projects will be amazing using this fluid and flexible layout, that is adaptive to all needs.